requirement.txt
burnysc2==6.4.0
firecrawl==1.3.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
JPype1==0.7.0
loguru==0.6.0
mpyq==0.2.5
multidict==6.1.0
nest-asyncio==1.6.0
numpy==1.24.4
openai==1.51.2
portpicker==1.6.0
propcache==0.2.0
protobuf==3.20.3
psutil==6.0.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
s2clientprotocol==5.0.13.92440.3
scipy==1.10.1
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
yarl==1.15.2
